net: pcs: implement Firmware node support for PCS driver
Implement the foundation of Firmware node support for PCS driver.

To support this, implement a simple Provider API where a PCS driver can
expose multiple PCS with an xlate .get function.

PCS driver will have to call fwnode_pcs_add_provider() and pass the firmware
node pointer and a xlate function to return the correct PCS for the
passed #pcs-cells.

This will register the PCS in a global list of providers so that
consumer can access it.

Consumer will then use fwnode_pcs_get() to get the actual PCS by passing the
firmware node pointer and the index for #pcs-cells.

For simple implementation where #pcs-cells is 0 and the PCS driver
expose a single PCS, the xlate function fwnode_pcs_simple_get() is
provided.

For advanced implementation a custom xlate function is required.

PCS driver on removal should first delete as a provider with
the usage of fwnode_pcs_del_provider() and then call
phylink_release_pcs() on every PCS the driver provides and

A generic function fwnode_phylink_pcs_parse() is provided for any MAC
driver that will declare PCS in DT (or ACPI).
This function will parse "pcs-handle" property and fill the passed array
with the parsed PCS. It's also possible to pass NULL as array to only
parse the PCS and return the number of the detected ones.

Co-developed-by: Daniel Golle <daniel@makrotopia.org>
Signed-off-by: Daniel Golle <daniel@makrotopia.org>
Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>